fix(tamagochi): populate metrics on state creation + ruff cleanups
Real bug: _get_or_create_state built TamagochiState without metric values,
relying on SQLAlchemy column defaults — but those apply at flush, leaving the
attributes None in-memory. The immediately-following _decay_in_place then did
`None - float` → TypeError on the very first read for a child. Set
energy/joy/learning/health=100 explicitly at construction.

Also: _clamp uses round() (round already returns int; RUF046) and ruff
auto-fixes (datetime.UTC alias, import sort) across the new modules.

Full suite 292 passed, ruff clean on all new files. This unbreaks the CI
that failed on the tamagochi/subscription push.

+++ duyo-backend/src/duyo/gamification/tamagochi.py
@@ -28,7 +28,7 @@ _METRICS = ("energy", "joy", "learning", "health")
 
 def _clamp(v: float) -> int:
     """Clamp to the valid 0-100 integer range (DUYO never dies → floor 0)."""
-    return max(0, min(100, int(round(v))))
+    return max(0, min(100, round(v)))
 
 
 @dataclass(frozen=True)
